Enforce active Group SAML SSO session
Prevents access to group resources when user hasn't signed in with SAML
and SSO enforcement is turned on.

Uses the session to track which SAML group the user has signed in with.
It then works by injecting the session into the user and then checking
this from the group policy.