Adds 'interruptible' column to ci_builds_metadata.
This attribute is added in this table in order to solve performance problems
that are caused if this attribute is set in the ci_builds options or directly
as a column in ci_builds.

This allow to lookup for cancellable pipelines directly through SQL queries instead
of having to iterate manually on all of them.