Implements pipeline cancellation based on the interruptible attribute.
A pipeline is cancelled when there is no uninterruptible build pending
or finished yet. A typical uninterruptible build would be a deployment step
whereas simple build/test steps could be considered as interruptible most
of the time.

This solves many uses cases, especially when the devs are only interested
in the last pipeline results. It can also save a lot of CPU time when
the pipeline builds are heavy and non-essential.