Expose previous design version images in GraphQL
As a DesignVersion represents a 'commit' of possibly several designs at
the same time, a DesignVersion has_many Designs.

In order to present an image for the DesignVersion we need to know the
corresponding Design also. In GraphQL this can be done by walking up
the tree until we find a DesignType node.

https://gitlab.com/gitlab-org/gitlab-ee/issues/9491