Can synchronously move repository storage
Add optional "sync" parameter to Project#change_repository_storage. This
functionality will not be exposed to the API, but is designed to be used
from Rails consoles by GitLab operators, in order to give more control
when migrating between storage shards.

Errors raised by ProjectUpdateRepositoryStorageWorker will be visible to
the operator, and can be handled synchronously in scripts that
orchestrate large numbers of repository migrations.

Also add an optional "skip_save" parameter, which allows calling code to
not have to manage saving the object itself. This is useful for rails
consoles and admin scripts. In production, this method is only called by
a service object which subsequently saves the project. The production
behaviour is unchanged by this addition.