Changed the way renaming works
In WebIDE, when re-naming an entry, instead of creating a new one,
update the existing one storing original values for eventual discarding.

After committing a re-named entry, remove the temporary parameters.

Handling files with spaces in path

Swap original entry with the re-named one in the state.openFiles if the
original entry has been opened

Mark re-named file with a 'renamed-' key

Fixed re-naming folders in WebIDE

Preserve changes to a re-named file

Changes and rename are marked as one entry in the "Commit" view