Added mutation to discard rename
If an entry was renamed previously but not really changed somehow else,
renaming it to the previous name dicards the renames all together and
removed the file from changed.

REVERT_RENAME_ENTRY mutation type added for easier discarding
functionality in the future

Added new help utility function to swap entries in parent tree with
sorting.