Apply patch to solve remaining issues
* Removes setting changed when renamed.
Other parts of the IDE are dependent that
this explicitly means the content has changed.
As a follow up to this, it updates the status
icon components to support prevPath.

* Resolves https://gitlab.com/gitlab-org/gitlab/merge_requests/16539#note_223326021

* Adds some if (parent) checks because now it's
possible for a file to have a parentPath which
doesn't point to an actual parent
(only if a nested rename is deleted).

* Removes entryPath from RENAME_ENTRY mutation.
The previous usage of this property was very confusing.
It was only used in special cases, and when it was it
was the same as path.

* Updates deleteEntry action to revert the rename if necessary.
Otherwise, we send the wrong path to the BE
(and some other less obvious issues).

* Updates deleteEntry to only stage if the file is not a tree
and not a new file (small fix to pre-existing issues).

* Moves the responsibility of adding/removing to "changedFiles"
or "stagedFiles" to the action renameEntry instead of the mutation
RENAME_ENTRY. This simplifies the mutation (and relevant helpers) a lot.

* Removes the mutation REVERT_RENAME_ENTRY and updates the mutation
RENAME_ENTRY to appropriately remove the prevPath properties if the
rename is resetting to the original name. This greatly simplifies the renameEntry action.

* Fix STAGE_FILE mutation by simplifying it's find staged file query.
We use this simple query elsewhere in the mutation, so it's safe to
keep it consistent.

* Adds some utils and spec.