WIP: Fix gitaly_address method
The Gitlab::GitalyClient::StorageSettings#gitaly_address currently
fails, because the GitLab configuration in production defines the
gitaly address entry for each node using a string key, and the method
actually tries to fetch the value based on a symbolic key.

Interestingly enough, the config/gitlab.yml.example uses a symbolic
key.

This fix simply introduces a secondary fetch using a string key, as
a fallback for the case in which a string key is used in the config.