Support skipping of ETag verification for object storage
Some S3 providers and transfer modes in Amazon S3 do not have ETags that
match the MD5 sum of the file. For these providers, we should provide a
way to skip this ETag verification until
https://gitlab.com/gitlab-org/gitlab-workhorse/issues/222 is
implemented.

This change sends the flag to Worhorse if the `skip_etag_verify` config
option is set. This flag will be used by
https://gitlab.com/gitlab-org/gitlab-workhorse/merge_requests/427

Relates to:
* https://gitlab.com/gitlab-org/gitlab-workhorse/issues/210
* https://gitlab.com/gitlab-org/charts/gitlab/issues/1564