Fix web_exporter shutdown
Currently, the `server.shutdown` will result
in shutdown of parent server, as internally it
uses `IO.pipe` to communicate, where the handle
is being passed between processes.

Closing sockets is exactly what is needed in
case of forks. We want the parent process to control
the server lifecycle, but we do not want to leak
sockets.