Allow sidekiq_exporter to fail on address in use
The https://gitlab.com/gitlab-org/gitlab/merge_requests/18575
make the exporters to fail hard, where previously it was allowed
to fail softly.

This retains the previous behavior and properly validates that
this is allowed.

This is needed due to lack of mechanism to support shared port
binding when running sidekiq-cluster mode.