Refactor pagination
This refactors the pagination helper and extracts the logic into
Gitlab::Pagination. The goal here is to change
Gitlab::Serializer::Pagination to *not* use Api::Helpers::Pagination
anymore. Instead, it calls the extracted classes directly.

This is a preparation to add more pagination strategies to the API only
(like keyset pagination).