Add handling for bad data in ci_build options
There’s a scenario by which a stringified ruby hash is making it into
`options` values, which can’t be serialized properly but is currently
not being handled. This fix resolves the scenarios we’re seeing where
the data is invalid, but we should take steps to not allow that bad
data into the system at all.