Fix missing RSS feed events
The ProjectsController currently loads a list of projects the user has
access to in the function load_projects. It uses this list to render the
actual list of projects, but it also uses this list to render the
activity feed in the atom form for the RSS feed.

The problem is that load_projects already paginates the list of projects
which leads to wrong results (only the first 20 projects are considered)
when rendering the RSS feed.

The error was intruduced in 837c3b3d, so this commit partially reverts
the changes made in that commit, moving the pagination back out of
load_projects and into the respective routes.