Add nodes to recursion analyzer IGNORED_FIELDS
As well as `field -> edges -> node`, we're able to structure queries
using the short `field -> nodes`:

```
{
  project(fullPath: "root/my-project") {
    id
    issues {
      nodes {
        discussions {
          nodes {
            notes {
              nodes {
                id
              }
            }
          }
        }
      }
    }
  }
}
```

The above query currently triggers the recursion analyzer to fail the
query, as it detects a recursion of `nodes`.

This commit adds `nodes` to the list of IGNORED_FIELDS of the recursion
analyzer.

It also adds a validation check that we do not use any IGNORED_FIELDS
in our field names.

https://gitlab.com/gitlab-org/gitlab/merge_requests/19596