Report command only notes correctly
This changes the behaviour of note creation to respond with HTTP-202
when the note is command only - i.e. no note was created, but the
changes mentioned were applied.

The response type is also changed to expose the command changes,
allowing the client to verify that their commands were correctly
applied, and allowing the client to distinguish from real notes that
have been persisted to the database from command only responses with a
mechanism other than just checking the note id.

Be aware that as a result of this change the response from this service
will now return notes that have a null ID, a condition that previously
excluded.