Runner Settings Update + Link to Docs
This is an update to the runner settings helper text. Before the text was to vague and didn't allow users to  understand or connect certain terms with their corresponding concepts. To solve this the text is now more clear and succinct. There's also a new link to the documentation that should help users to quickly gain context on the concept of Runner.