Add author restriction push rule
The committer restriction push rule can be used to ensure that every
commit that is pushed was commited by the same user. However, the
author may be any other user. This could be misleading if the author
was maliciously set to another user.

This new push rule allows project maintainers and instance admins to
require the commit author email to be the same as the user pushing the
commit in the same way that the committer restriction requires the
commit comitter email to be the same as the user pushing the commit.