Add QA test for prometheus alerts
The configuration of alerts depends on a lot of moving parts and has
been prone to issue. This test sets up a cluster, creates a project,
connects the cluster to the project, sets up Auto Devops, deploys the
application, then performs basic CRUD operations on alerts from the
metrics dashboard. Alerts are configurated by updating Prometheus, so
a future test could build upon the assertions by ensuring that the alert
can indeed be fired.