Move deployment state hooks to ActiveRecord hooks
When creating a new deployment using the API, we bypass all the state
machine code in the Deployment model. For example, if we create a
deployment with the status of "success", we would never set
"finished_at" as this attribute is set by a state machine transition.

In this commit we move two state machine hooks to ActiveRecord hooks:

1. The hook used for setting "finished_at".
2. The hook for scheduling the Deployments::FinishedWorker worker.

This ensures that deployments created using the API and CI both trigger
the same workers, hooks, etc.