Start pipeline on Jenkins push hook
When we trigger the webhook, there is no way for users to see the
status on the commit or MR before Jenkins updates the status.
To fix this, we can start a new pipeline and set the status to
"pending". If the webhook fails, the pipeline will fail as well.
It is up to Jenkins to pick up the pipeline and change the status
depending on the build status later