Gitlab::Geo::Replicator implementing existing supported events
A Replicator is the central class that is responsible to publish an event
in a primary node and process it on a secondary one, using contextual metadata
to help reflect the operation from one side to the other.

There is a 1:1 relationship between a Replicator and it's associated model.

Optionally when we want to replicate an event that is not tied to a model,
like purging a cache, we can omit this relation.