Cleanup passing down of avatar source
To cleanup the way the user avatar gets passed down
it was extracted into a seperate computed prop.

It also clearly returns undefined as `null` will not be caught
by a Vue Type check

Alongside I added specs to solidify the code.
Sadly the scope is all global which is something that needs
to be addressed in a seperate RFC