Prune epics management end-to-end test
This test fails intermittently on CI in the assertion after editing
the epic title, as it can be seen in the below job:

https://gitlab.com/gitlab-org/gitlab-qa/-/jobs/387863693

Editing and deletion of epics have already a good coverage in lower
level tests, as it can be seen from the following files:

- ee/spec/features/epics/update_epic_spec.rb
- ee/spec/features/epics/delete_epic_spec.rb

With that, it's safe to remove these steps from the end-to-end test
suite and still be sure that these functionalities are still being
tested.

Besides that, now the test is focused on only one thing.

Finally, this change reduces the code in 17 lines.