Create optimal indexes for created_at order
This is for the projects API. Note that we can't have one index only
because we cannot scan the index backwards for both `sort=asc,sort=desc`
options. That is because the tie breaker is always `id desc` and hence
the sort order is not suitable for a backward scan for one of the
options.

Relates to https://gitlab.com/gitlab-org/gitlab/issues/36390