Possible fix for flaky sidebar specs
In both EE and FOSS specs, two labels are clicked in quick succession.
When the test fails, it appears that the first label was never actually
clicked.

To prevent some timing issue, we now pause after the first click to
verify the right number of active labels are present in the dropdown.

Closes https://gitlab.com/gitlab-org/gitlab/issues/118659