Improve error tracking of merge failures
We saw in https://gitlab.com/gitlab-org/gitaly/issues/2395 that a
failure to create a worktree while attempting to merge a squashed commit
would log an error only via `Rails.logger`, which made it hard to tell
what happened and whether the error was a systemic problem.

We improve this by using the error tracking interface, which not only
logs the exception to Sentry but also to `exceptions_json.log`.

This does log some acceptable errors (e.g. `Branch has been updated
since the merge was requested`), but it makes it possible to see how
often that problem is occurring.