Simplify project milestone resource
The fabricate method is not in use, and so it just adds unnecessary
complexity. This is why it is being removed.
The same applies to the description attribute.

Also, instead of having a title= method, better having it initiali-
zed with a default value so that it doesn't need to be provided
when creating the resource in the test file.

Finally, the test that uses the updated resource was updated to use
the default title.