Implement `ndjson` support for `import/export`
This implements `ndjson` and streaming `json`
support to handle two cases:

- big `project.json` (legacy way)
- new `.ndjson` format, where each relation
  receives a separate file, and each item is
  stored per-line

This can properly detect old and a new file contents,
without any changes to the files, and by maintaining
backward compatibility.

This implements a trick to support streaming json writer
to append data additively.

This overall when `exporting legacy/ndjson` or `importing ndjson`
allows us to have a constant memory for the process,
and also significantly reduces latency of the data processing
due to not escaping to the native.