Replace file type check with content whitelist
File type check results in user inconvenience so we moved to a content
whitelist check to complement existing extension checks.

Created a ContentTypeWhitelist to bring forward the whitelist feature
available in Carrierwave 2 but not in the current 1.3.1 version.

Applied the ContentTypeWhitelist to the GitlabUploader base class.

DesignV432x230Uploader specs confirm patch improvements function where
they previously did not.