Update max_seats_used counter for GL.com subscriptions in real time
Instead of waiting for the daily script that updates the max_seats_used
counter we're updating it from every place where Group or Project
members are created or updated.

Calculating the total number of billable members can be an expensive
task for large groups like gitlab-org. That's why we're using a feature flag, we need to
disable this feature in case we observe some performance issues.