Remove unnecessary checking in git controller
At the moment we peform a check to ensure that a
guest user has access to the repository. This
check is performed again in `GitAccess`.

We don't need to perform this first check.

Besides, there is a bug with this check
because it only checks `download_code`. This
check can have a bad behavior when the object
of the git operation is a wiki.