Refactor resizeable-panel from CollapsibleSidebar
This is an incremental refactoring to support a collapsible left sidebar
in the Web IDE.

We want to make the resizeable panel not need to be the responsibility
of CollapsibleSidebar anymore.

This restructure required some styling changes across elements,
especially due to the mix of Flexbox and non-Flexbox in some child
pane components.

It also contains some minor fixes:

* related to removing tooltips and focus highlighting when clicking
  on the buttons.

* Removing header slots which will not be needed by the left side,
  and switching the single remaining slot for the footer to a
  default instead of a named slot.