Migration to create vulnerabilities
This migration creates vulnerabilities records for
vulnerability_occurrences records that exist on the
latest successful pipeline of the default branch of
a given porject and do not already have a
vulnerabilities record created.
It also updates the states of vulnerabilities records
to dismissed if the corresponding
vulnerability_occurrences record was dismissed.