Fix bug committing content when snippet created
When the snippet is created we commit the content
to the recently created repository. Nevertheless,
these actions are inside a db transaction and
the commit action triggers a request
handled by a different thread.

This means that in the thread the snippet
does not exists and, therefore, the operation
fails.