Add wiki_events feature flag
This ensures that the methods we use for retrieving events respect the
:wiki_events feature flag.

This flag must be applied in two places - both in the event_collection
and in the events_finder, since there exist places in the code base
where the event_collection is used without the finder (for example in
the projects or the groups controllers) so it must be applied there, but
in the events_finder there is a branch that by-passes the
event_collection, so we also need it in the finder.

This flag is off by default.