Move snippet feature flag check to controllers
At the moment, class `GitAccessSnippet` ensures that
the feature flag :version_snippets is enabled in order
to allow/disallow snippet git actions.

Nevertheless, there are git operations, like when
we update the snippet info, that we want to perform
always, no matter the feature flag status.

We need to move the check outside the access class into
the controllers and the internal API.