Allow selecting all queues with sidekiq-cluster
Because it was designed for selecting specific queues, sidekiq-cluster
didn't have a convenient method to select all queues. As we want this to
become the default sidekiq entry point for all GitLab installations, and
the default is to handle all queues in a single process, we need that
option.

(This was possible before: using `--experimental-queue-selector`, you
could ask for something like `has_external_dependencies=true,false`, but
that's not obvious.)

Instead of making `*` a wildcard anywhere, which would be more general,
this commit just allows a single `*` in a queue group to represent all
queues. While this is less general, it's also simpler to implement, and
we can assume that YAGNI for the wildcards.