Geo Replicable - Generalize Mutations
This MR is an effort at MVC.

This MR splits off from:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/26410

The overall focus of this and following related
changes is to replace Geo_Designs with
Geo_Replicable.

This MR focuses on the Mutations and
private Store functions.

Following MRs:
Replication Type and API
Public Store Functions and State
Vue Components and Filenames