Backfill LfsObjectsProject records of forks
To fix the behavior of existing forks, LfsObjectsProject records
need to be backfilled. This is so we can remove the need to depend
on the source when looking for LFS objects from forks.

This is based on implementation in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/25343 that
was reverted. The difference is the query used for finding the
projects to be migrated is now performant.