Support sending traffic to redis-cache read replica
This is a proof of concept for improving the scalability of
redis-cache by sending traffic to read replicas.

There are some redis keys that may not be safe for stale reads.
Thus, the idea here is to perform stale reads selectively, we
can incrementally migrate calls that are deemed safe to replicas.
That can be done by passing `stale_ok: true` to Rails.cache read
calls.

We perform a relatively crude load balancing across replicas:
the sentinel support in redis-rb will select a random healthy
replica at connection time.