Separate actions on authorized keys from gitlab-shell
The gitlab-shell project used to contain code to manage authorized_keys
but this was long ago merged into the main gitlab project. However, we
still indirect through `Gitlab::Shell` for all key actions. This is a
barrier to understanding and makes things unnecessarily complex. Since
we want to remove `Gitlab::Shell` in the near(ish) future, it also gets
in the way.

This commit removes that indirection and introduces a separate worker
to handle add_key / remove_key calls while maintaining backward
compatibility with in-flight sidekiq jobs.