Stop using Gitlab::Shell#create_repository in specs
This method is going away, and its use in specs hinders that effort.
In specs, it is used to set up scenarios where a repository conflicts
with an attempt to create one, so going through a ::Repository is
typically not an option. We can use `TestEnv#create_bare_repository`
instead, though.