Add --report-in-issues to use issues to track tests
Allows test results in JUnit XML files to be used to create/update
GitLab issues to report the status of tests.

Add unit tests

Add gitlab client as a runtime dependency

Limit title to 255 characters

Search for issues by the full title

Also don't raise an exception if there's a problem finding a specific
test. Just output a warning.

Search in quotes for exact match

Include test full description in issue description
So search can find the right issue when the test name is too long to be
unique

Filter by exact title
This avoids matching on a test with a name that's a substring of another