Add an accessibility MR widget
Refactor the test reports component to be able
to report any kind of test suite report. Since the
accessibility reports are going to have the same
data structure, we should just reuse the existing
component. Perhaps we should move the folder
into `vue_shared`, but I want to make sure this
works first.