Create accessibility report component
My last assumption about being able to use the
`grouped_test_reports_app` component was
wrong. That is used for multiple reports and we
only have one that we are rendering. Also, the
data structure is slightly different.

So I reverted my previous changes, and instead
created a new store and app for the a11y report.

Some of this will need to be changed based on
more information from backend (data structure).