Iteration 3: rename first
We rename column first, and then create a view
that allows to access the renamed column
using old name.

This adapts all indexes, which is desired from
the application perspective: we want to be able
to continue inserting data for time being,
and ensure that application is fully stable
after we restart.