Docs: try out suggestion-level substitutions
Our current substitutions file in Vale is set to the error level,
which is not appropriate for all of the suggestions we want to make
to the technical writing team. This commit adds another substition
file with a name that makes clear the items within it are at a
suggestion level, rather than an error level.