Ensure Global Search metrics collected on one node
Right now we're trying to add monitoring in
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/2046 but
unfortunately our metrics are being collected randomly on different
sidekiq nodes which makes it difficult to plot the data sensibly. This
change will ensure it's only running on one node. Since this is a global
metric that counts a ZSET in Redis it really doesn't matter which node
it runs on but it makes sense to run on the dedicated elasticsearch
sidekiq node.