Make setting the hook environment unconditional
This one-line change deserves a little explanation. Prior to
b94246278843477f9a9e05097e162300df76bd8c, the hook env was set
unconditionally. In that commit, the environment was transformed to
switch from relative to absolute paths for object directories, etc.

In 52bfd064627c6a63b82bc81e7ebabac299b25eed we stopped doing that
transformation, but the `if project` line was moved from the transform
to the (previously unconditional) hook env set line. It didn't cause
problems, because every repository had a project.

In time, we introduced repositories for personal snippets, where this
is no longer true - and they broke. So we transformed the check to read
"if container" in 36ea513552d697a8c32821eb0c3550a25f6f0f05, because we
weren't sure if it was necessary or not, and we were in a hurry.

Now that I've dug into the history of this change, I'm confident we can
make this call unconditional, so this commit does that.