Use Sampler instead of worker for monitoring Global Search queue
Introduced in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/27384 we are
tracking the size of the buffer queue used for incremental Elasticsearch
updates. This was implemented as a worker that runs every minute but
this has some problems as it runs on random sidekiq nodes and causes the
data in Prometheus to appear as multiple unpredictable lines and there
is not way to easily know which is the most up to date metric. For this
reason it makes more sense to run it predictably and also on all sidekiq
nodes so that even if one is down we'd still see the queue size growing.