danger: Commits can start with lowercase letters
Git, but also other projects, require commit messages to start with the
section in which the changes was made. This is helpful when running for
example `git log` on a directory, as by just reading the first word of
the commit subject, a lot of commits do not require reading the rest of
the message.

By removing the danger rule, GitLab-Rails would allow for this too. And
as such could later reap the benefits.