Remove the Blob#container abstraction
Blobs do not *necessarily* have a project. Personal snippets and group
wikis have repositories, with blobs, that are not part of any project.

The "container" abstraction was introduced to hide this difference, but
it proves difficult to use well. Since the contexts that want to use
`::Blob` typically *already* know what the container is, we can get
cleaner code by removing the ability to look up a project/container
from the blob itself.