Fix spec failure caused by memoization
Repository#exists? is memoized, so we have to clear it if we want a
broken repo that responds to exists? correctly. This change allows us to
only every have a single instance of ProjectWiki per project,
simplifying specs a lot.

A new trait, `:broken_wiki_repo` is added to abstract the state
twiddling.