Streamline scaling and availability documentation
This commit creates a new docs page called System Architecture,
nested under the Install section of the left nav bar. The info
that was previously on the Scaling and High Availability pages
has been deduplicated and refactored into the new System
Architecture page, which required moving about a dozen files.