Introduce CommandStore and ExecutionResponse
This introduces two new abstractions: CommandStore and
ExecutionResponse:

CommandStore abstracts the command registry.
This moves command registration (and iteration/lookup) to a new class,
CommandStore, a singleton command registry.

ExecutionResponse abstracts the result of running quick action commands.
The response is responsible for knowing which commands have been
applied, what the message and warnings are, and what updates need to be
applied.

Command application is moved away from an instance method, since no
instance variables need to be known.