Update project_daily_statistics synchronally
The ProjectDailyStatisticsWorker queue accumulates jobs
from time to time, considering how often the git_upload_pack
endpoint is called (taking up to 7 minutes to take the job
from the queue in a few cases).

Turns out that even though this is a low-priority queue (and
it's OK to take a bit more time to process it) it was
seen that the work it does takes around 1.8s (99th percentile),
and 0.8s (99th percentile) DB time, which doesn't explain
involving Sidekiq (and Redis) in the work pipeline.

Here we move the processing out of Sidekiq and back to the
web request behind a feature flag (and deprecate the worker).