Find audit events based on audit level
Introduce new `Audit::Level` class to facilitate finding based on audit
level. This object can be later used by `AuditLogFinder` to construct
ActiveRecord query.

For `Audit::GroupLevel`, we would like to see all Project-level events
related to projects that belong to the group. By default, if the
`resource` argument is not a known type, we will return
`Audit::InstanceLevel` and later inferred as `AuditEvent.all`
scope.