Allow subclasses of BasePipeline to resolved
Currently, one can only pass Symbols to `Pipeline.[]` which is
convenient in most cases assuming that the pipeline is namespaced to
`Banzai::Pipeline`.

In some cases, we want a namespace a pipeline to a specific domain
namespace (e.g `StatusPage::PostProcessPipeline`). In those cases we had
to pass a underscored version of the namespaced class name (e.g.
`:'_status_page::_post_process_pipeline'` or
`:'status_page/post_process_pipeline'`).

This commit allows passing a subclass of
`Banzai::Pipeline::BasePipeline` to `Pipeline.[]`.