Skip orchestrated tests in the Praefect jobs
The Praefect jobs are intended to run all the unorchestated tests
against an instance using Praefect for storage. So although it is
technically an orchestrated scenario, it's orchestration that
allows the regular tests to run, but *not* those tests that
need additional orchestration provided by other scenarios.

One of those tests is `move_project_create_fork_spec.rb`, which
is intended to be run in the RepositoryStorage scenario that
configures additional storage. So the test fails in the Praefect
scenario because Praefect is configured as the default storage,
not additional.

This also removes the parallel param from the quarantine
jobs, because there's typically not enough of them to need
to run in multiple jobs.