Tutorial revisions from the TW team
This MR provides revisions from the TW team that eliminate prose
linting issues, reduce the file size of the accompanying image, and
bring the wording closer to the GitLab style guide. A lot of
paragraphs were combined to get rid of single-sentence paragraphs,
and phrasing was changed to eliminate passive voice and future tense.