Update Geo Uploads Filter Nav
This MR is work towards the final product
proposed here:
https://gitlab.com/gitlab-org/gitlab/-/issues/213219

This MR updates the Geo Uploads
filters and top nav to match the design spec.

To do so only required swapping the old nav for
the new nav that is saved inside a re-usable
partial.

For consistancy purposes until Uploads
are added to Geo Self Service, the
never status was mapped to pending.