Handle conflict no CI and pipelines must succeed
When there are no CI services found and
the pipeline must succeed option is turned
on, this is considered an invalid usecase
since the pipeline will never be successful.
Because of this, we add a meaningful message
to the UI that explain this situation and link
to our documentation while in the Merge request.