Improve docs about certificates when enabling TLS for Gitaly nodes
In order to configure Gitaly to communicate via TLS, its certificate
needs to be installed to all Gitaly nodes including itself as well as
all off its clients. The paragraph mentioning this is confusingly
worded, making it easy to miss its intention.

Reword the paragraph to make its intention clearer.