Retry all system call errors and parsing errors
There are many many system call errors. It may be more efficient to
retry on any of them rather than waiting for a failure to occur before
adding them to the list.

We've seen a increasing number of parsing errors that happen when
searching for an issue. That should only happen if GitLab returned
an error as HTML, which should only happen if something has gone
very wrong. So we retry those errors too because they're likely to
be transient.